requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.33.12
botocore==1.33.12
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
defusedxml==0.8.0rc2
distro==1.8.0
Django==5.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
jmespath==1.0.1
kombu==5.3.4
oauthlib==3.2.2
openai==1.3.9
prompt-toolkit==3.0.43
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pynamodb==5.5.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.8.2
six==1.16.0
sniffio==1.3.0
social-auth-core==4.5.1
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
